+++ ubo_app/services/000-rgb-ring/rgb_ring_client.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 
 import socket
 from pathlib import Path
+from typing import Sequence
 
+from ubo_app.constants import SOCKET_PATH
 from ubo_app.logging import logger
 from ubo_app.store import dispatch
 from ubo_app.store.services.rgb_ring import RgbRingSetIsConnectedAction
-
-LM_SOCKET_PATH = Path('/run/ubo').joinpath('ledmanagersocket.sock').as_posix()
 
 
 class RgbRingClient:
@@ -28,21 +28,22 @@ class RgbRingClient:
 
     def __init__(self: RgbRingClient) -> None:
         self.server_socket: socket.SocketType | None = None
-        if Path(LM_SOCKET_PATH).exists():
+        if Path(SOCKET_PATH).exists():
             self.server_socket = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
             try:
-                self.server_socket.connect(LM_SOCKET_PATH)
-                dispatch(RgbRingSetIsConnectedAction(is_connected=True))
+                self.server_socket.connect(SOCKET_PATH)
             except Exception as exception:  # noqa: BLE001
                 dispatch(RgbRingSetIsConnectedAction(is_connected=False))
                 logger.error('Unable to connect to the socket', exc_info=exception)
                 return
+            else:
+                dispatch(RgbRingSetIsConnectedAction(is_connected=True))
 
     def __del__(self: RgbRingClient) -> None:
         if self.server_socket is not None:
             self.server_socket.close()
 
-    def send(self: RgbRingClient, cmd: str) -> None:
+    def send(self: RgbRingClient, cmd: Sequence[str]) -> None:
         if not self.server_socket:
             return
-        self.server_socket.send(cmd.encode('utf-8'))
+        self.server_socket.send(' '.join(['led', *cmd]).encode('utf-8'))